fix linter findings for windows and darwin
